rename StringConstant to StrEnum
The StringConstant is really much like the enum.StrEnum or the
strenum.StrEnum. It makes the code a bit easier to read for others if
the same term (StrEnum) is used also here.
