+++ wakepy/core/__init__.py
@@ -9,7 +9,7 @@ from .activationresult import StageName as StageName
 from .calls import Call as Call
 from .calls import DbusMethodCall as DbusMethodCall
 from .configuration import CURRENT_SYSTEM as CURRENT_SYSTEM
-from .constant import StringConstant as StringConstant
+from .strenum import StrEnum as StrEnum
 from .dbus import BusType as BusType
 from .dbus import DbusAddress as DbusAddress
 from .dbus import DbusMethod as DbusMethod
